+++ yash-env/CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to `yash-env` will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.9.0] - Unreleased
+
+- External dependency versions:
+    - yash-syntax 0.15.0 → 0.16.0
+- Internal dependency versions:
+    - annotate-snippets 0.11.4 → 0.12.3
+
 ## [0.8.1] - Unreleased
 
 ### Changed
@@ -471,6 +478,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial implementation of the `yash-env` crate
 
+[0.9.0]: https://github.com/magicant/yash-rs/releases/tag/yash-env-0.9.0
 [0.8.1]: https://github.com/magicant/yash-rs/releases/tag/yash-env-0.8.1
 [0.8.0]: https://github.com/magicant/yash-rs/releases/tag/yash-env-0.8.0
 [0.7.1]: https://github.com/magicant/yash-rs/releases/tag/yash-env-0.7.1